refactor: saveAuth, removeAuth 삭제 및 TokenStage와 AuthStorage 통합
#141
